ENG-01: extract _preprocessing (MCUVScaler, center, scale)
Also add an explicit __all__ to _common so CodeQL does not flag the
DataMatrix type alias (referenced only in lazy annotation strings) as an
unused global. _pca_pls re-imports the moved names.
